feat: The schema of the coze API interface comparison table is completed
