+++ _coze-plugin/tests/test_akshare_api.rs
@@ -1,5 +1,6 @@
 use std::{fs::File, io::Write, sync::LazyLock, time::Duration};
 
+use data_mind::schema;
 use reqwest::{Client, ClientBuilder};
 use serde_json::Value;
 
@@ -61,6 +62,136 @@ async fn test_stock_news_main_cx() {
 
     println!("res len = {}", res.len());
     let mut file = File::create("../tmp/财经内容精选.json").unwrap();
+    file.write_all(serde_json::to_string_pretty(&res).unwrap().as_bytes())
+        .unwrap();
+}
+
+#[actix_web::test]
+async fn test_stock_rank_cxg_ths() {
+    let symbols = ["创月新高", "半年新高", "一年新高", "历史新高"];
+
+    for symbol in symbols {
+        let res: Vec<schema::akshare::StockRankCxgThs> = TEST_HTTP_CLIENT
+            .get(with_base_url("/stock_rank_cxg_ths"))
+            .query(&[("symbol", symbol)])
+            .send()
+            .await
+            .unwrap()
+            .error_for_status()
+            .unwrap()
+            .json()
+            .await
+            .unwrap();
+
+        println!("symbol = {symbol}, res len = {}", res.len());
+
+        let mut file = File::create(format!("../tmp/技术指标-创新高-{}.json", symbol)).unwrap();
+        file.write_all(serde_json::to_string_pretty(&res).unwrap().as_bytes())
+            .unwrap();
+    }
+}
+
+#[actix_web::test]
+async fn test_stock_rank_cxd_ths() {
+    let symbols = ["创月新低", "半年新低", "一年新低", "历史新低"];
+
+    for symbol in symbols {
+        let res: Vec<schema::akshare::StockRankCxdThs> = TEST_HTTP_CLIENT
+            .get(with_base_url("/stock_rank_cxd_ths"))
+            .query(&[("symbol", symbol)])
+            .send()
+            .await
+            .unwrap()
+            .error_for_status()
+            .unwrap()
+            .json()
+            .await
+            .unwrap();
+
+        println!("symbol = {symbol}, res len = {}", res.len());
+
+        let mut file = File::create(format!("../tmp/技术指标-创新低-{}.json", symbol)).unwrap();
+        file.write_all(serde_json::to_string_pretty(&res).unwrap().as_bytes())
+            .unwrap();
+    }
+}
+
+#[actix_web::test]
+async fn test_stock_rank_lxsz_ths() {
+    let res: Vec<schema::akshare::StockRankLxszThs> = TEST_HTTP_CLIENT
+        .get(with_base_url("/stock_rank_lxsz_ths"))
+        .send()
+        .await
+        .unwrap()
+        .error_for_status()
+        .unwrap()
+        .json()
+        .await
+        .unwrap();
+
+    println!("res len = {}", res.len());
+
+    let mut file = File::create("../tmp/技术指标-连续上涨.json").unwrap();
+    file.write_all(serde_json::to_string_pretty(&res).unwrap().as_bytes())
+        .unwrap();
+}
+
+#[actix_web::test]
+async fn test_stock_rank_lxxd_ths() {
+    let res: Vec<schema::akshare::StockRankLxxdThs> = TEST_HTTP_CLIENT
+        .get(with_base_url("/stock_rank_lxxd_ths"))
+        .send()
+        .await
+        .unwrap()
+        .error_for_status()
+        .unwrap()
+        .json()
+        .await
+        .unwrap();
+
+    println!("res len = {}", res.len());
+
+    let mut file = File::create("../tmp/技术指标-连续下跌.json").unwrap();
+    file.write_all(serde_json::to_string_pretty(&res).unwrap().as_bytes())
+        .unwrap();
+}
+
+#[actix_web::test]
+async fn test_stock_rank_cxfl_ths() {
+    let res: Vec<schema::akshare::StockRankCxflThs> = TEST_HTTP_CLIENT
+        .get(with_base_url("/stock_rank_cxfl_ths"))
+        .send()
+        .await
+        .unwrap()
+        .error_for_status()
+        .unwrap()
+        .json()
+        .await
+        .unwrap();
+
+    println!("res len = {}", res.len());
+
+    let mut file = File::create("../tmp/技术指标-持续放量.json").unwrap();
+    file.write_all(serde_json::to_string_pretty(&res).unwrap().as_bytes())
+        .unwrap();
+}
+
+#[actix_web::test]
+async fn test_stock_rank_cxsl_ths() {
+    let res: Vec<schema::akshare::StockRankCxslThs> = TEST_HTTP_CLIENT
+        .get(with_base_url("/stock_rank_cxsl_ths"))
+        .send()
+        .await
+        .unwrap()
+        .error_for_status()
+        .unwrap()
+        .json()
+        .await
+        .unwrap();
+
+    println!("res len = {}", res.len());
+
+    let mut file = File::create("../tmp/技术指标-持续缩量.json").unwrap();
     file.write_all(serde_json::to_string_pretty(&res).unwrap().as_bytes())
         .unwrap();
 }
