+++ src/schema/akshare/mod.rs
@@ -1,3 +1,6 @@
+//! 仅仅用于从akshare api之中获取数据，然后插入到clickhouse之中的一些
+//! json schema
+
 pub mod a_index;
 pub mod a_stock;
