+++ src/schema/mod.rs
@@ -1,2 +1,2 @@
 pub mod akshare;
-pub mod coze;
+pub mod service;
